refactor: setup new colors
